Convert setup.py to pyproject.toml
Remove the setup.py based packaging and replace it with pyproject.toml.

The used build system is still setuptools.  Editable installs need
setuptools >= 64, which requires Python >= 3.7.  This means that the
package can't be built with Python 3.6 anymore, but if it is still
installable from a wheel which can be built with newer Python
version (e.g. by using `pipx run build --wheel`).

Remove the click<8 constraint, because it seems that 8 works too.

Add passari-workflow to dependencies, since it's actually required.

Rename the extra named "sphinx" to "docs", because that is the name
suggested in Sphinx documentation.

Update README on how to build the docs, since the old way of using
`python setup.py build_sphinx` has been deprecated for a while and
doesn't work anymore with the new pyproject.toml.