Allow setting CUSTOM_LOG_LEVELS for the flask app
Add a configuration option named CUSTOM_LOG_LEVELS and use it in the
Flask app initialization (create_app) to configure log level
customizations for the loggers named in that option.

This allows e.g. to override the log level for passlib library, which
seems to otherwise pollute the terminal with some surplus messages if
general log level is set to DEBUG (10).