tests: Simplify database pytest fixture
Stop using the DatabaseJanitor from pytest-postgresql to start a
separate PostgreSQL server for the tests, but instead connect to
existing database server via the configured database settings and
appending "_test" to the database name.  If there is no configuration
(nor DATABASE_URL env var), fallback to connecting to passari_test
database on local unix socket.

Use functions from sqlalchemy-utils to drop the test database if it
already exists and to (re-)create it to ensure that a fresh test
database is available for each test case.

This makes the database fixture much simpler and allows running the
tests on host that doesn't have PostgreSQL installed (by connecting to
remote PostgreSQL instance).