Improve database initialization (init-db/create-db)
Rename the create-db command to init-db, because it's not actually
creating a database but just initializing it for Web UI usage.  However,
the create-db command is still provided for backward compatibility.

Also make the command check if all tables already exist and print
informational "Database already initialized" message in that case.